Wrote tests for sets backend logic

+++ shared/schemas.ts
@@ -14,7 +14,7 @@ export const ExerciseNameSchema = z.coerce.string().trim().min(1, "Exercise name
 
 export const UpdateSetSchema = z.object({
     weight_kg: z.coerce.number().max(9999.999).nonnegative(),
-    reps: z.coerce.number().min(1),
+    reps: z.coerce.number().int().min(1),
 }).partial()
 
 export type WorkoutProps = {
